refactor(instrument): reorg utility functions
